split tests some more
